Conflicts:
	drivers/misc/mediatek/imgsensor/src/mt6735m/camera_hw/kd_camera_hw.c
	drivers/misc/mediatek/include/mt-plat/mt6735/include/mach/mt_battery_meter.h
	drivers/misc/mediatek/video/mt6735/dispsys/ddp_dsi.c
	drivers/misc/mediatek/video/mt6735/videox/mt6735m/disp_drv_platform.h
	drivers/misc/mediatek/video/mt6735/videox/primary_display.c

+++ drivers/misc/mediatek/connectivity/common/combo/core/wmt_ic_6630.c
@@ -1961,7 +1961,7 @@ done:
 static INT32 wmt_stp_wifi_lte_coex(VOID)
 {
 	INT32 iRet;
-	UINT32 addr;
+	ULONG addr;
 	WMT_GEN_CONF *pWmtGenConf;
 
 	/*Get wmt config */
